Delegate backend insert/view creation to Ibis natives
- Move default insert() to IbisBackend: validates columns, reorders, and
  calls connection.insert(). DuckDB and Spark now use the default; SQLite
  keeps its raw-cursor override so inserts honor our BEGIN/ROLLBACK
  transaction state (ibis.sqlite.insert auto-commits).
- Drop SparkBackend's temp-view-and-SQL insert, _prepare_data_for_spark,
  and _prepare_write_data: ibis.pyspark.create_table / insert already
  convert tz-aware pandas to UTC-naive and accept pyarrow directly.
- Simplify create_view_from_parquet on DuckDB and Spark to
  connection.create_view(name, connection.read_parquet(path)).
- Hoist shared _validate_insert_columns / _get_columns / _select_columns
  / _row_count helpers from the per-backend files into base.py.
